feat #9 : OpenAI 서비스 인터페이스 및 dto 분리
