+++ src/main/java/com/example/study_monster_back/openAi/dto/response/OpenAiStudyFeedbackResponse.java
@@ -1,0 +1,11 @@
+package com.example.study_monster_back.openAi.dto.response;
+
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@NoArgsConstructor
+public class OpenAiStudyFeedbackResponse {
+    private String feedback;
+    private String futureLearningStrategy;
+}
