+++ src/test/java/com/example/study_monster_back/openAi/service/OpenAiServiceImplTest.java
@@ -1,6 +1,6 @@
 package com.example.study_monster_back.openAi.service;
 
-import com.example.study_monster_back.openAi.service.OpenAiService.StudyFeedbackResponse;
+import com.example.study_monster_back.openAi.dto.response.OpenAiStudyFeedbackResponse;
 import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -11,9 +11,9 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 @SpringBootTest
 @DisplayName("openAiService 로 ")
-class OpenAiServiceTest {
+class OpenAiServiceImplTest {
     @Autowired
-    private OpenAiService openAiService;
+    private OpenAiServiceImpl openAiServiceImpl;
 
     @Disabled
     @Test
@@ -24,7 +24,7 @@ class OpenAiServiceTest {
         String userMessage = "hello who are you?";
 
         // when
-        String response = openAiService.getPlainAnswer(systemMessage, userMessage);
+        String response = openAiServiceImpl.getPlainAnswer(systemMessage, userMessage);
 
         // then
         System.out.println(response);
@@ -46,14 +46,14 @@ class OpenAiServiceTest {
 
             두 번째 주차부터는 실제 프로젝트를 진행하면서 디자인 감각을 키우는 시간을 가졌습니다. 개인 기록장 게시판부터 시작해서, 스터디 그룹 모집 페이지까지 다양한 웹 서비스의 UI를 직접 디자인해 보았습니다.
             처음에는 깔끔한 디자인을 위해 무엇을 빼야 할지에 집중했는데, 강사님이 "좋은 디자인은 기능을 돋보이게 하는 것"이라고 말씀해주신 것이 기억에 남습니다. 사용자가 원하는 정보를 쉽게 찾을 수 있도록 시각적 계층 구조를 만드는 연습을 많이 했어요.
- 
+             
             프로토타이핑의 힘
             마지막 주차에서는 프로토타이핑에 대해 배웠습니다. 정적인 디자인에 상호작용을 추가하니 마치 실제 웹사이트처럼 작동하는 모습을 보며 큰 성취감을 느꼈습니다.
             특히 상세 페이지의 탭 인터페이스나 모달 창 같은 복잡한 인터랙션도 피그마 내에서 구현할 수 있다는 점이 인상적이었어요. 개발자와 소통할 때도 이렇게 구체적인 프로토타입이 있으면 훨씬 수월할 것 같습니다.
             """;
 
         // when
-        StudyFeedbackResponse studyFeedback = openAiService.getStudyFeedback(title, article);
+        OpenAiStudyFeedbackResponse studyFeedback = openAiServiceImpl.getStudyFeedback(title, article);
 
         // then
         System.out.println(studyFeedback.getFeedback());
